update ci and remove unecessary deps
