+++ README.md
@@ -1,13 +1,1 @@
-#### Next.js
-
-#### TypeScript
-
-#### Zustand (state management)
-
-#### TailwindCSS
-
-#### Vitest
-
-#### ESLint
-
-#### Prettier
+# WIP
