+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './button'
 export * from './message'
+export * from './textbox'
